search function V1
i implemented a primative search function that searches the phones based on the keywords submitted by the search bar from the user. the only issue is that the name of the phone is still difficult to find like "galaxy" because the name of the phone can be of two words, same goes for the brand name such as "One Plus". will have to look further into an efficient solution to this problem.